+++ AClient.cs
@@ -35,7 +35,7 @@ namespace GrenciCPA
 
         //misc
         private bool active;
-        private double balance;
+        private decimal balance;
         private string notes;
         private List<string> characteristics;
         private string preferred;
@@ -65,7 +65,7 @@ namespace GrenciCPA
         public string Phone { get => phone; set => phone = value; }
         public string Email { get => email; set => email = value; }
         public bool Active { get => active; set => active = value; }
-        public double Balance { get => balance; set => balance = value; }
+        public decimal Balance { get => balance; set => balance = value; }
         public string Notes { get => notes; set => notes = value; }
         public List<string> Characteristics { get => characteristics; set => characteristics = value; }
         public string Preferred { get => preferred; set => preferred = value; }
